+++ src/main/gate/creole/annic/apache/lucene/store/OutputStream.java
@@ -112,7 +112,7 @@ public abstract class OutputStream {
        throws IOException {
     final int end = start + length;
     for (int i = start; i < end; i++) {
-      final int code = (int)s.charAt(i);
+      final int code = s.charAt(i);
       if (code >= 0x01 && code <= 0x7F)
 	writeByte((byte)code);
       else if (((code >= 0x80) && (code <= 0x7FF)) || code == 0) {
